New Changes 1/20 HAYDEN
Changed a bunch. Now all integer values are stored in order in the
ArrayList<Integer> changeValues. This will allow us to then use these
values easily as we turn them into RPMs. We also need to make it so
that the changes of RPM do something, and that the simulated quadcopter
doesn’t fly out of frame after like ten ticks.