New Code From Morning 1/21
Now reads file, stores it all in array, and reads these values to
determine next tick. Also uses images instead of a square (we may need
to enlarge image). Still, the values are kind of hard to determine and
it seems kind of random how it does it. Also the line is rewritten over
to be from the origin (well, 100, 100). We need it to save the last
point and make the new line from the point, not deleting the old line.